pyproject.toml
[project]
name = "sentiment-analysis"
version = "0.2.0"
description = "IMDB/Rotten Tomatoes sentiment analysis with transformer models"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "datasets==2.14.4",
    "ipython==8.18.1",
    "jupyter==1.1.1",
    "matplotlib==3.9.4",
    "nltk==3.9.1",
    "pandas==2.3.2",
    "pyyaml==6.0.2",
    "scikit-learn==1.6.1",
    "torch==2.5.0",
    "torchvision==0.20.0",
    "tqdm==4.67.1",
    "transformers[torch]==4.46.0",
    "click==8.1.7",
    "numpy==1.26.4",
]

[project.scripts]
sentiment-train = "main:cli"
